@@ -22,23 +22,23 @@
   },
   "homepage": "https://github.com/bjyoungblood/hapi-serve-s3#readme",
   "dependencies": {
-    "aws-sdk": "^2.33.0",
-    "boom": "^4.3.1",
-    "content": "^3.0.3",
+    "aws-sdk": "^2.165.0",
+    "boom": "^5.2.0",
+    "content": "^3.0.6",
     "content-disposition": "^0.5.2",
-    "hoek": "^4.1.0",
-    "joi": "^10.3.1",
-    "uuid": "^3.0.1"
+    "hoek": "^4.2.0",
+    "joi": "^12.0.0",
+    "uuid": "^3.1.0"
   },
   "devDependencies": {
-    "eslint": "^3.18.0",
-    "eslint-config-airbnb-base": "^11.1.2",
-    "eslint-plugin-import": "^2.2.0",
+    "eslint": "^4.12.1",
+    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint-plugin-import": "^2.8.0",
     "expect": "^1.20.2",
-    "form-data": "^2.1.2",
-    "hapi": "^16.1.0",
-    "mocha": "^3.2.0",
-    "rimraf": "^2.6.1",
+    "form-data": "^2.3.1",
+    "hapi": "^16.6.2",
+    "mocha": "^4.0.1",
+    "rimraf": "^2.6.2",
     "s3rver": "^1.0.3"
   },
   "engines": {
